requirements.txt
PyPDF2>=3.0.0
stripe
pytesseract>=0.3.10
Pillow>=9.0.0
opencv-python-headless>=4.5.0
numpy>=1.21.0
requests>=2.25.0
qrcode[pil]>=7.4.2 